dcp-712 Move IngestSubmitter test suite to its own file
